Add files via upload
A simple Java program to simulate a ball undergoing projectile  motion.  The ball is launched from the lower left hand side of the screen and follows the typical parabolic trajectory.  When the ball  hits the ground, we use conservation of energy to determine a new  initial velocity assuming a constant energy loss per bounce.  From  here we can repeat the simulation until a point is reached where the ball stops bouncing.